[241] Move properties-related services to form-related projects
Bug: https://github.com/eclipse-sirius/sirius-components/issues/241
