+++ backend/sirius-web-spring-collaborative-forms/src/main/java/org/eclipse/sirius/web/spring/collaborative/forms/PropertiesEventProcessorFactory.java
@@ -23,12 +23,12 @@ import org.eclipse.sirius.web.collaborative.api.services.ISubscriptionManagerFac
 import org.eclipse.sirius.web.collaborative.forms.api.IFormEventHandler;
 import org.eclipse.sirius.web.collaborative.forms.api.IFormEventProcessor;
 import org.eclipse.sirius.web.collaborative.forms.api.IPropertiesDefaultDescriptionProvider;
+import org.eclipse.sirius.web.collaborative.forms.api.IPropertiesDescriptionService;
 import org.eclipse.sirius.web.collaborative.forms.api.IWidgetSubscriptionManagerFactory;
 import org.eclipse.sirius.web.collaborative.forms.api.PropertiesConfiguration;
 import org.eclipse.sirius.web.core.api.IEditingContext;
 import org.eclipse.sirius.web.core.api.IObjectService;
 import org.eclipse.sirius.web.forms.description.FormDescription;
-import org.eclipse.sirius.web.services.api.representations.IPropertiesDescriptionProvider;
 import org.springframework.stereotype.Service;
 
 /**
@@ -39,7 +39,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class PropertiesEventProcessorFactory implements IRepresentationEventProcessorFactory {
 
-    private final IPropertiesDescriptionProvider propertiesDescriptionProvider;
+    private final IPropertiesDescriptionService propertiesDescriptionService;
 
     private final IPropertiesDefaultDescriptionProvider propertiesDefaultDescriptionProvider;
 
@@ -51,10 +51,10 @@ public class PropertiesEventProcessorFactory implements IRepresentationEventProc
 
     private final IWidgetSubscriptionManagerFactory widgetSubscriptionManagerFactory;
 
-    public PropertiesEventProcessorFactory(IPropertiesDescriptionProvider propertiesDescriptionProvider, IPropertiesDefaultDescriptionProvider propertiesDefaultDescriptionProvider,
+    public PropertiesEventProcessorFactory(IPropertiesDescriptionService propertiesDescriptionService, IPropertiesDefaultDescriptionProvider propertiesDefaultDescriptionProvider,
             IObjectService objectService, List<IFormEventHandler> formEventHandlers, ISubscriptionManagerFactory subscriptionManagerFactory,
             IWidgetSubscriptionManagerFactory widgetSubscriptionManagerFactory) {
-        this.propertiesDescriptionProvider = Objects.requireNonNull(propertiesDescriptionProvider);
+        this.propertiesDescriptionService = Objects.requireNonNull(propertiesDescriptionService);
         this.propertiesDefaultDescriptionProvider = Objects.requireNonNull(propertiesDefaultDescriptionProvider);
         this.objectService = Objects.requireNonNull(objectService);
         this.formEventHandlers = Objects.requireNonNull(formEventHandlers);
@@ -73,7 +73,7 @@ public class PropertiesEventProcessorFactory implements IRepresentationEventProc
         if (IFormEventProcessor.class.isAssignableFrom(representationEventProcessorClass) && configuration instanceof PropertiesConfiguration) {
             PropertiesConfiguration propertiesConfiguration = (PropertiesConfiguration) configuration;
 
-            List<FormDescription> formDescriptions = this.propertiesDescriptionProvider.getPropertiesDescriptions();
+            List<FormDescription> formDescriptions = this.propertiesDescriptionService.getPropertiesDescriptions();
             Optional<Object> optionalObject = this.objectService.getObject(editingContext, propertiesConfiguration.getObjectId());
             if (optionalObject.isPresent()) {
                 Object object = optionalObject.get();
